requirements.txt
Django==2.0.1
requests
djangorestframework